requirements.txt
boto3==1.24.38
botocore==1.27.38
jmespath==1.0.1
python-crontab==2.6.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.11
